\name{as.gt}
\alias{as.gt}
\title{Extract genotype profile for a single contributor from rGTs output}
\description{Extract genotype profile for a single contributor from rGTs output}
\usage{
as.gt(res,ind)
}
\arguments{
  \item{res}{Output from rGTs}
  \item{ind}{Integer, which individual's genotype profile should be extracted}
}
\value{Data frame, genotype profile for selected individual, for format see \code{\link{KinMix-package}}.}
\author{
Peter Green (P.J.Green@bristol.ac.uk)
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
