\name{intoMix}
\alias{intoMix}
\title{Edit output from rGTs to omit individuals with NA amounts of DNA}
\description{Edit output from rGTs to omit individuals with NA amounts of DNA}
\usage{
intoMix(res)
}
\arguments{
  \item{res}{Output from \code{rGTs}}
}
\value{The edited data structure}
\author{
Peter Green (P.J.Green@bristol.ac.uk)
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
